Auto-install overleaf packages from list (#9)
The overleaf image includes a minimal set of latex packages by default.
They advise you install the required packages and save that docker
image, but that process would need to be repeated when upgrading.

Instead, this entrypoint automatically installs the required packages on
every container start. This is not efficient for a real deployment, or
for a large number of packages, but is a simple and effective solution
for small deployments such as this.

Fixes #7.